+++ src/main/java/com/prography/lighton/auth/application/port/AuthVerificationService.java
@@ -1,4 +1,4 @@
-package com.prography.lighton.auth.application;
+package com.prography.lighton.auth.application.port;
 
 public interface AuthVerificationService {
