+++ src/main/java/com/prography/lighton/auth/security/filter/JwtAuthenticationFilter.java
@@ -4,8 +4,8 @@ import static com.prography.lighton.common.constant.JwtConstants.AUTHORIZATION_H
 import static com.prography.lighton.common.constant.JwtConstants.BEARER_PREFIX;
 import static com.prography.lighton.common.constant.JwtConstants.ROLE_PREFIX;
 
-import com.prography.lighton.auth.application.TokenProvider;
 import com.prography.lighton.auth.application.exception.InvalidTokenException;
+import com.prography.lighton.auth.application.port.TokenProvider;
 import com.prography.lighton.common.constant.SecurityWhitelist;
 import io.jsonwebtoken.Claims;
 import jakarta.servlet.FilterChain;
