+++ src/test/java/com/prography/lighton/common/config/DisableSecurityConfig.java
@@ -2,7 +2,7 @@ package com.prography.lighton.common.config;
 
 import static org.mockito.Mockito.mock;
 
-import com.prography.lighton.auth.application.TokenProvider;
+import com.prography.lighton.auth.application.port.TokenProvider;
 import org.springframework.boot.test.context.TestConfiguration;
 import org.springframework.context.annotation.Bean;
 import org.springframework.security.config.annotation.web.builders.HttpSecurity;
